requirements.txt
langchain==0.1.14
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
python-dotenv
flask
gitdb==4.0.11
GitPython==3.1.40
smmap==5.0.1
langchain_elasticsearch
